Expose package.json version in UI
