+++ CHANGELOG.md
@@ -1,0 +1,15 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+## Unreleased changes
+
+### Added
+- Expose package.json version in UI
+
+### Changed
+
+### Fixed
+- Fix overflow issue, making some UI elements to display text vertically
+
+## [1.0.0] - (2021-05-20)
