+++ package.json
@@ -94,7 +94,7 @@
     "@types/enzyme": "^3.10.5",
     "@types/express": "^4.17.7",
     "@types/jest": "^26.0.8",
-    "@types/node": "^14.0.27",
+    "@types/node": "^14.17.0",
     "@types/ws": "^7.2.6",
     "babel-loader": "^8.1.0",
     "concurrently": "^5.2.0",
@@ -102,7 +102,6 @@
     "css-loader": "^4.2.0",
     "cssnano": "^4.1.10",
     "elm": "^0.19.1-3",
-    "elm-hot-webpack-loader": "^1.1.7",
     "elm-webpack-loader": "^8.0.0",
     "enzyme": "^3.11.0",
     "enzyme-to-json": "^3.5.0",
