+++ tsconfig.json
@@ -8,7 +8,6 @@
     "outDir": "./dist",
     "lib": ["es2015", "dom"],
     "types": [],
-    "jsx": "react"
   },
-  "files": ["./types.d.ts", "./src/client/client.tsx"]
+  "files": ["./types.d.ts", "src/client/client.ts"]
 }
